refactor: applications, 오류 핸들러 분리
